add protected and public routes in homejs; moved barchartjs to /maps
