+++ client/src/components/maps/BarChart.js
